feat: migrate to keycloak (#117)
* feat: start supporting AI4OS Keycloak

* feat: start using roles

* feat: update permissions check

* feat: move to AI4EOSC Vault

* feat: check audiences in auth

* fix: use roles instead of groups

For now both return the same values but this can change in the future.

* fix: allow llm for demo users

* fix: multi-vo fetch

* feat: adapt batch methods

+++ ai4papi/routers/v1/batch.py
@@ -46,13 +46,11 @@ def get_deployments(
 
     # If no VOs, then retrieve jobs from all user VOs
     # Always remove VOs that do not belong to the project
-    if not vos:
-        vos = auth_info["vos"]
     vos = set(vos).intersection(set(papiconf.MAIN_CONF["auth"]["VO"]))
     if not vos:
         raise HTTPException(
             status_code=401,
-            detail=f"The provided Virtual Organizations do not match with any of your available VOs: {auth_info['vos']}.",
+            detail=f"Your VOs do not match available VOs: {papiconf.MAIN_CONF['auth']['VO']}.",
         )
 
     user_jobs = []
@@ -116,7 +114,7 @@ def get_deployment(
 
     # Retrieve authenticated user info
     auth_info = auth.get_user_info(token=authorization.credentials)
-    auth.check_vo_membership(vo, auth_info["vos"])
+    auth.check_authorization(auth_info, vo)
 
     # Retrieve the associated namespace to that VO
     namespace = papiconf.MAIN_CONF["nomad"]["namespaces"][vo]
@@ -171,7 +169,7 @@ def create_deployment(
     """
     # Retrieve authenticated user info
     auth_info = auth.get_user_info(token=authorization.credentials)
-    auth.check_vo_membership(vo, auth_info["vos"])
+    auth.check_authorization(auth_info, vo)
 
     # Load module configuration
     # To avoid duplicating too much code, we use the standard job deployment
@@ -422,7 +420,7 @@ def delete_deployment(
     """
     # Retrieve authenticated user info
     auth_info = auth.get_user_info(token=authorization.credentials)
-    auth.check_vo_membership(vo, auth_info["vos"])
+    auth.check_authorization(auth_info, vo)
 
     # Delete deployment
     r = nomad.delete_deployment(
